fix: restore main architecture ratchet
