+++ server/src/routes/chats.ts
@@ -9,8 +9,6 @@ import {
   createChatQueuedMessageSchema,
   createSideChatSchema,
   forkChatConversationSchema,
-  parseCodexInlineVisualDirectives,
-  parseRudderInlineVisualPlacements,
   steerChatQueuedMessageSchema,
   updateChatConversationSchema,
   updateChatQueuedMessageSchema,
@@ -18,10 +16,8 @@ import {
   type ChatContextLink,
   type ChatControlDisposition,
   type ChatConversation,
-  type ChatInlineVisualMapping,
   type ChatMessage,
   type ChatQueueRequestActor,
-  type RudderInlineVisualMapping,
 } from "@rudderhq/shared";
 import { Router, type Request, type Response } from "express";
 import multer from "multer";
@@ -85,6 +81,7 @@ import {
   type ChatBackgroundTimer,
 } from "./chat-background-runtime.js";
 import { wakeIssueAssigneeAfterChatConversion } from "./chat-issue-assignment-wakeup.js";
+import { attachGeneratedChatFiles } from "./chats.generated-attachments.js";
 import {
   isMultipartRequest,
   paginateChatMessages,
@@ -648,179 +645,16 @@ export function chatRoutes(
   ) {
     const createdMessages: ChatMessage[] = [];
     const { chatTurnId, turnVariant } = turnContext;
-    const attachGeneratedFiles = async (message: ChatMessage, generatedAttachments: ChatGeneratedAttachment[] | undefined) => {
-      const finalDirectives = parseCodexInlineVisualDirectives(assistantReply.body).directives;
-      const finalPlacements = parseRudderInlineVisualPlacements(assistantReply.body).placements;
-      const inlineVisuals = (assistantReply.inlineVisuals ?? []).filter((visual) =>
-        finalDirectives.some((directive) =>
-          directive.index === visual.directiveIndex && directive.file === visual.file
-        )
-      );
-      const inlineVisualsV1 = (assistantReply.inlineVisualsV1 ?? []).filter((visual) =>
-        finalPlacements.some((placement) => placement.slot === visual.slot)
-      );
-      const generatedFiles = (generatedAttachments ?? []).filter((generated) =>
-        (generated.source !== "codex_inline_visual" && generated.source !== "rudder_inline_visual")
-        || (generated.source === "codex_inline_visual" && inlineVisuals.some((visual) =>
-          visual.status === "captured"
-          && visual.directiveIndex === generated.directiveIndex
-          && visual.file === generated.directiveFile
-        ))
-        || (generated.source === "rudder_inline_visual" && inlineVisualsV1.some((visual) =>
-          visual.status === "captured"
-          && visual.slot === generated.slot
-          && visual.file === generated.originalFilename
-        ))
-      );
-      if (generatedFiles.length === 0 && inlineVisuals.length === 0 && inlineVisualsV1.length === 0) return message;
-      const attachments: ChatAttachment[] = [];
-      const attachmentByVisualIndex = new Map<number, ChatAttachment>();
-      const attachmentByVisualSlot = new Map<number, ChatAttachment>();
-      const createdAttachmentRecords: Array<{ attachmentId: string; orgId: string; objectKey: string }> = [];
-      try {
-        for (const generated of generatedFiles) {
-          if (generated.body.length > MAX_ATTACHMENT_BYTES) {
-            throw new ChatAssistantStreamError(
-              `Generated attachment exceeds ${MAX_ATTACHMENT_BYTES} bytes`,
-              assistantReply.body,
-              generatedFiles,
-              { partialBodyUserVisible: true },
-            );
-          }
-          const stored = await storage.putFile({
-            orgId: conversation.orgId,
-            namespace: `chats/${conversation.id}/generated`,
-            originalFilename: generated.originalFilename,
-            contentType: generated.contentType,
-            body: generated.body,
-          });
-          let attachment;
-          try {
-            attachment = await svc.createAttachment({
-              orgId: conversation.orgId,
-              conversationId: conversation.id,
-              messageId: message.id,
-              provider: stored.provider,
-              objectKey: stored.objectKey,
-              contentType: stored.contentType,
-              byteSize: stored.byteSize,
-              sha256: stored.sha256,
-              originalFilename: stored.originalFilename,
-              createdByAgentId: replyingAgentId,
-              createdByUserId: null,
-            });
-          } catch (error) {
-            await storage.deleteObject(conversation.orgId, stored.objectKey).catch((cleanupError) => {
-              logger.warn({ err: cleanupError, objectKey: stored.objectKey }, "failed to clean up generated chat object");
-            });
-            throw error;
-          }
-          createdAttachmentRecords.push({
-            attachmentId: attachment.id,
-            orgId: conversation.orgId,
-            objectKey: stored.objectKey,
-          });
-          const typedAttachment = attachment as ChatAttachment;
-          const publicAttachment = generated.source === "codex_inline_visual" || generated.source === "rudder_inline_visual"
-            ? (({ provider: _provider, objectKey: _objectKey, ...safe }) => safe)(typedAttachment)
-            : typedAttachment;
-          attachments.push(publicAttachment as ChatAttachment);
-          if (generated.source === "codex_inline_visual") {
-            attachmentByVisualIndex.set(generated.directiveIndex, publicAttachment as ChatAttachment);
-          } else if (generated.source === "rudder_inline_visual") {
-            attachmentByVisualSlot.set(generated.slot, publicAttachment as ChatAttachment);
-          }
-        }
-      let structuredPayload = message.structuredPayload ?? null;
-      let persistedLegacyMappings: ChatInlineVisualMapping[] = [];
-      let persistedRudderMappings: RudderInlineVisualMapping[] = [];
-      if (inlineVisuals.length > 0) {
-        persistedLegacyMappings = inlineVisuals.map((visual) => {
-          if (visual.status === "captured") {
-            const attachment = attachmentByVisualIndex.get(visual.directiveIndex);
-            if (attachment) {
-              return {
-                directiveIndex: visual.directiveIndex,
-                file: visual.file,
-                status: "ready" as const,
-                attachmentId: attachment.id,
-              };
-            }
-          }
-          return {
-            directiveIndex: visual.directiveIndex,
-            file: visual.file,
-            status: "unavailable" as const,
-            reason: visual.status === "unavailable" ? visual.reason : "capture_failed",
-          };
-        });
-        structuredPayload = {
-          ...(structuredPayload ?? {}),
-          inlineVisuals: persistedLegacyMappings,
-        };
-      }
-      if (inlineVisualsV1.length > 0) {
-        persistedRudderMappings = inlineVisualsV1.map((visual) => {
-          if (visual.status === "captured") {
-            const attachment = attachmentByVisualSlot.get(visual.slot);
-            if (attachment) {
-              return {
-                version: 1 as const,
-                slot: visual.slot,
-                file: visual.file,
-                status: "ready" as const,
-                attachmentId: attachment.id,
-                contentType: "text/html" as const,
-                byteSize: attachment.byteSize,
-                sha256: attachment.sha256,
-              };
-            }
-          }
-          return {
-            version: 1 as const,
-            slot: visual.slot,
-            file: visual.file,
-            status: "unavailable" as const,
-            reason: visual.status === "unavailable" ? visual.reason : "capture_failed",
-          };
-        });
-        structuredPayload = {
-          ...(structuredPayload ?? {}),
-          inlineVisualsV1: persistedRudderMappings,
-        };
-      }
-        if (persistedLegacyMappings.length > 0 || persistedRudderMappings.length > 0) {
-          const internallyUpdated = await svc.updateMessageInternalInlineVisuals(
-            conversation.id,
-            message.id,
-            {
-              ...(persistedLegacyMappings.length > 0 ? { inlineVisuals: persistedLegacyMappings } : {}),
-              ...(persistedRudderMappings.length > 0 ? { inlineVisualsV1: persistedRudderMappings } : {}),
-            },
-          );
-          if (!internallyUpdated) throw new Error("Failed to persist trusted inline visual mapping");
-          structuredPayload = internallyUpdated.structuredPayload ?? structuredPayload;
-        }
-        return {
-          ...message,
-          structuredPayload,
-          attachments: [...(message.attachments ?? []), ...attachments],
-        } as ChatMessage;
-      } catch (error) {
-        for (const created of createdAttachmentRecords.reverse()) {
-          const removed = await svc.removeAttachment(created.attachmentId).catch((cleanupError) => {
-            logger.warn({ err: cleanupError, attachmentId: created.attachmentId }, "failed to remove generated chat attachment");
-            return null;
-          });
-          if (removed?.assetDeleted) {
-            await storage.deleteObject(created.orgId, created.objectKey).catch((cleanupError) => {
-              logger.warn({ err: cleanupError, objectKey: created.objectKey }, "failed to remove generated chat object");
-            });
-          }
-        }
-        throw error;
-      }
-    };
+    const attachGeneratedFiles = (message: ChatMessage, generatedAttachments: ChatGeneratedAttachment[] | undefined) =>
+      attachGeneratedChatFiles({
+        assistantReply,
+        generatedAttachments,
+        message,
+        conversation,
+        replyingAgentId,
+        storage,
+        chats: svc,
+      });
     const saveAssistantMessage = async (input: {
       kind: "message" | "ask_user" | "issue_proposal" | "operation_proposal";
       body: string;
