+++ server/src/services/chats.ts
@@ -18,7 +18,7 @@ import {
   messengerCustomGroups,
   organizations
 } from "@rudderhq/db";
-import { chatInlineVisualMappingsFromStructuredPayload, MESSENGER_FORK_GROUP_DEFAULT_ICON, parseCodexInlineVisualDirectives, parseRudderInlineVisualPlacements, rudderInlineVisualMappingsFromStructuredPayload, sanitizeChatStructuredPayload, type ChatControlDisposition, type ChatInlineVisualMapping, type ChatProviderControlDisposition, type ChatQueuedMessage, type ChatQueuedMessagePayload, type ChatQueuedMessageStatus, type ChatQueueRequestActor, type ChatStreamTranscriptEntry, type RudderInlineVisualMapping } from "@rudderhq/shared";
+import { MESSENGER_FORK_GROUP_DEFAULT_ICON, sanitizeChatStructuredPayload, type ChatControlDisposition, type ChatInlineVisualMapping, type ChatProviderControlDisposition, type ChatQueuedMessage, type ChatQueuedMessagePayload, type ChatQueuedMessageStatus, type ChatQueueRequestActor, type ChatStreamTranscriptEntry, type RudderInlineVisualMapping } from "@rudderhq/shared";
 import { and, asc, desc, eq, gt, gte, inArray, isNotNull, isNull, lt, or, sql } from "drizzle-orm";
 import { randomUUID } from "node:crypto";
 import { conflict, notFound, unprocessable } from "../errors.js";
@@ -57,6 +57,11 @@ import {
   withOperationProposalDecisionState,
   withPersistedTranscript,
 } from "./chats.helpers.js";
+import {
+  copyForkInlineVisualMessages,
+  listRecentUserChatMessages,
+  updateTrustedInlineVisualMappings,
+} from "./chats.inline-visual-persistence.js";
 import type { ChatServerQueueClaim, ConversationSourceMetadata, ConversationSummaryCursor, MessageHydrationRow } from "./chats.types.js";
 import { issueApprovalService } from "./issue-approvals.js";
 import { issueService } from "./issues.js";
@@ -3171,126 +3176,13 @@ export function chatService(db: Db) {
       const forkMessages = input.sourceMessageId
         ? messagesToCopy.slice(0, messagesToCopy.findIndex((message) => message.id === input.sourceMessageId) + 1)
         : messagesToCopy;
-      if (forkMessages.length > 0) {
-        for (const message of forkMessages) {
-          const copiedMessageId = randomUUID();
-          const visualMappings = chatInlineVisualMappingsFromStructuredPayload(message.structuredPayload);
-          const visualMappingsV1 = rudderInlineVisualMappingsFromStructuredPayload(message.structuredPayload);
-          const readyAttachmentIds = [
-            ...visualMappings,
-            ...visualMappingsV1,
-          ].filter((mapping) => mapping.status === "ready").map((mapping) => mapping.attachmentId);
-          const copiedAttachmentIdBySourceId = new Map<string, string>();
-          const pendingVisualAttachments: Array<{
-            id: string;
-            orgId: string;
-            conversationId: string;
-            messageId: string;
-            assetId: string;
-          }> = [];
-          if (readyAttachmentIds.length > 0) {
-            const sourceVisualAttachments = await tx
-              .select({
-                id: chatAttachments.id,
-                assetId: chatAttachments.assetId,
-                contentType: assets.contentType,
-                byteSize: assets.byteSize,
-                sha256: assets.sha256,
-                originalFilename: assets.originalFilename,
-                createdByAgentId: assets.createdByAgentId,
-                createdByUserId: assets.createdByUserId,
-              })
-              .from(chatAttachments)
-              .innerJoin(assets, eq(chatAttachments.assetId, assets.id))
-              .where(and(
-                eq(chatAttachments.conversationId, source.id),
-                eq(chatAttachments.messageId, message.id),
-                inArray(chatAttachments.id, readyAttachmentIds),
-              ));
-            const safeVisualAttachments = sourceVisualAttachments.filter((attachment) => {
-              if (
-                attachment.contentType !== "text/html"
-                || !attachment.createdByAgentId
-                || attachment.createdByUserId
-              ) return false;
-              return visualMappings.some((mapping) =>
-                mapping.status === "ready"
-                && mapping.attachmentId === attachment.id
-                && mapping.file === attachment.originalFilename
-              ) || visualMappingsV1.some((mapping) =>
-                mapping.status === "ready"
-                && mapping.attachmentId === attachment.id
-                && mapping.file === attachment.originalFilename
-                && mapping.contentType === attachment.contentType
-                && mapping.byteSize === attachment.byteSize
-                && mapping.sha256 === attachment.sha256.toLowerCase()
-              );
-            });
-            if (safeVisualAttachments.length > 0) {
-              pendingVisualAttachments.push(...safeVisualAttachments.map((attachment) => {
-                const copiedAttachmentId = randomUUID();
-                copiedAttachmentIdBySourceId.set(attachment.id, copiedAttachmentId);
-                return {
-                  id: copiedAttachmentId,
-                  orgId: input.orgId,
-                  conversationId: child.id,
-                  messageId: copiedMessageId,
-                  assetId: attachment.assetId,
-                };
-              }));
-            }
-          }
-          const copiedVisualMappings = visualMappings.map((mapping) => {
-            if (mapping.status === "unavailable") return mapping;
-            const copiedAttachmentId = copiedAttachmentIdBySourceId.get(mapping.attachmentId);
-            return copiedAttachmentId
-              ? { ...mapping, attachmentId: copiedAttachmentId }
-              : {
-                directiveIndex: mapping.directiveIndex,
-                file: mapping.file,
-                status: "unavailable" as const,
-                reason: "fork_source_missing",
-              };
-          });
-          const copiedVisualMappingsV1 = visualMappingsV1.map((mapping) => {
-            if (mapping.status === "unavailable") return mapping;
-            const copiedAttachmentId = copiedAttachmentIdBySourceId.get(mapping.attachmentId);
-            return copiedAttachmentId
-              ? { ...mapping, attachmentId: copiedAttachmentId }
-              : {
-                version: 1 as const,
-                slot: mapping.slot,
-                file: mapping.file,
-                status: "unavailable" as const,
-                reason: "fork_source_missing",
-              };
-          });
-          const copiedStructuredPayload = {
-            ...(copiedVisualMappings.length > 0 ? { inlineVisuals: copiedVisualMappings } : {}),
-            ...(copiedVisualMappingsV1.length > 0 ? { inlineVisualsV1: copiedVisualMappingsV1 } : {}),
-          };
-          await tx.insert(chatMessages).values({
-            id: copiedMessageId,
-            orgId: input.orgId,
-            conversationId: child.id,
-            role: message.role,
-            kind: message.kind,
-            status: message.status === "streaming" ? "interrupted" : message.status,
-            body: message.body,
-            structuredPayload: Object.keys(copiedStructuredPayload).length > 0 ? copiedStructuredPayload : null,
-            approvalId: null,
-            runId: null,
-            replyingAgentId: message.replyingAgentId,
-            chatTurnId: null,
-            turnVariant: 0,
-            createdAt: message.createdAt,
-            updatedAt: message.updatedAt,
-          });
-          if (pendingVisualAttachments.length > 0) {
-            await tx.insert(chatAttachments).values(pendingVisualAttachments);
-          }
-        }
-      }
+      await copyForkInlineVisualMessages({
+        tx,
+        messages: forkMessages,
+        sourceConversationId: source.id,
+        targetConversationId: child.id,
+        orgId: input.orgId,
+      });
 
       const [systemEvent] = await tx
         .insert(chatMessages)
@@ -3559,32 +3451,7 @@ export function chatService(db: Db) {
   }
 
   async function listRecentUserMessages(conversationId: string, limit: number) {
-      const ecmascriptTrimWhitespace = "\u0009\u000A\u000B\u000C\u000D\u0020\u00A0\u1680\u2000\u2001\u2002\u2003\u2004\u2005\u2006\u2007\u2008\u2009\u200A\u2028\u2029\u202F\u205F\u3000\uFEFF";
-      const boundedLimit = Math.min(5, Math.max(1, Math.trunc(limit)));
-      const conversationOrgIds = db
-        .select({ orgId: chatConversations.orgId })
-        .from(chatConversations)
-        .where(eq(chatConversations.id, conversationId));
-      const rows = await db
-        .select({
-          id: chatMessages.id,
-          role: chatMessages.role,
-          kind: chatMessages.kind,
-          body: chatMessages.body,
-          createdAt: chatMessages.createdAt,
-        })
-        .from(chatMessages)
-        .where(and(
-          eq(chatMessages.conversationId, conversationId),
-          inArray(chatMessages.orgId, conversationOrgIds),
-          isNull(chatMessages.supersededAt),
-          eq(chatMessages.role, "user"),
-          eq(chatMessages.kind, "message"),
-          sql<boolean>`btrim(${chatMessages.body}, ${ecmascriptTrimWhitespace}) <> ''`,
-        ))
-        .orderBy(desc(chatMessages.createdAt), desc(chatMessages.id))
-        .limit(boundedLimit);
-      return rows.reverse();
+      return listRecentUserChatMessages(db, conversationId, limit);
   }
 
   async function listMessages(conversationId: string, options: { includeTranscript?: boolean } = {}) {
@@ -3905,96 +3772,13 @@ export function chatService(db: Db) {
       inlineVisualsV1?: RudderInlineVisualMapping[];
     },
   ) {
-    const existing = await db
-      .select()
-      .from(chatMessages)
-      .where(and(eq(chatMessages.conversationId, conversationId), eq(chatMessages.id, messageId)))
-      .then((rows) => rows[0] ?? null);
-    if (!existing || existing.role !== "assistant" || existing.status !== "completed") return null;
-
-    const legacyDirectives = parseCodexInlineVisualDirectives(existing.body).directives;
-    const rudderPlacements = parseRudderInlineVisualPlacements(existing.body).placements;
-    const requestedLegacy = chatInlineVisualMappingsFromStructuredPayload({ inlineVisuals: input.inlineVisuals ?? [] })
-      .filter((mapping) => legacyDirectives.some((directive) =>
-        directive.index === mapping.directiveIndex && directive.file === mapping.file
-      ));
-    const requestedRudder = rudderInlineVisualMappingsFromStructuredPayload({ inlineVisualsV1: input.inlineVisualsV1 ?? [] })
-      .filter((mapping) => rudderPlacements.some((placement) => placement.slot === mapping.slot));
-    const readyAttachmentIds = [...requestedLegacy, ...requestedRudder]
-      .filter((mapping) => mapping.status === "ready")
-      .map((mapping) => mapping.attachmentId);
-    const attachmentRows = readyAttachmentIds.length === 0 ? [] : await db
-      .select({
-        id: chatAttachments.id,
-        contentType: assets.contentType,
-        byteSize: assets.byteSize,
-        sha256: assets.sha256,
-        originalFilename: assets.originalFilename,
-        createdByAgentId: assets.createdByAgentId,
-        createdByUserId: assets.createdByUserId,
-      })
-      .from(chatAttachments)
-      .innerJoin(assets, eq(chatAttachments.assetId, assets.id))
-      .where(and(
-        eq(chatAttachments.orgId, existing.orgId),
-        eq(chatAttachments.conversationId, conversationId),
-        eq(chatAttachments.messageId, messageId),
-        inArray(chatAttachments.id, readyAttachmentIds),
-      ));
-    const inlineVisuals = requestedLegacy.map((mapping): ChatInlineVisualMapping => {
-      if (mapping.status === "unavailable") return mapping;
-      const attachment = attachmentRows.find((row) =>
-        row.id === mapping.attachmentId
-        && row.contentType === "text/html"
-        && row.originalFilename === mapping.file
-        && row.createdByAgentId === existing.replyingAgentId
-        && !row.createdByUserId
-        && row.byteSize > 0
-        && row.byteSize <= 2 * 1024 * 1024
-      );
-      return attachment ? mapping : {
-        directiveIndex: mapping.directiveIndex,
-        file: mapping.file,
-        status: "unavailable",
-        reason: "ownership_mismatch",
-      };
+    return updateTrustedInlineVisualMappings({
+      db,
+      hydrateMessages,
+      conversationId,
+      messageId,
+      ...input,
     });
-    const inlineVisualsV1 = requestedRudder.map((mapping): RudderInlineVisualMapping => {
-      if (mapping.status === "unavailable") return mapping;
-      const attachment = attachmentRows.find((row) =>
-        row.id === mapping.attachmentId
-        && row.contentType === mapping.contentType
-        && row.originalFilename === mapping.file
-        && row.byteSize === mapping.byteSize
-        && row.sha256.toLowerCase() === mapping.sha256
-        && row.createdByAgentId === existing.replyingAgentId
-        && !row.createdByUserId
-      );
-      return attachment ? mapping : {
-        version: 1,
-        slot: mapping.slot,
-        file: mapping.file,
-        status: "unavailable",
-        reason: "ownership_mismatch",
-      };
-    });
-    const base = sanitizeChatStructuredPayload(existing.structuredPayload) ?? {};
-    const next = {
-      ...base,
-      ...(inlineVisuals.length > 0 ? { inlineVisuals } : {}),
-      ...(inlineVisualsV1.length > 0 ? { inlineVisualsV1 } : {}),
-    };
-    const [updated] = await db
-      .update(chatMessages)
-      .set({ structuredPayload: next, updatedAt: new Date() })
-      .where(and(
-        eq(chatMessages.orgId, existing.orgId),
-        eq(chatMessages.conversationId, conversationId),
-        eq(chatMessages.id, messageId),
-      ))
-      .returning();
-    const [hydrated] = await hydrateMessages([updated]);
-    return hydrated ?? null;
   }
 
   async function markInterruptedStreamingMessages(conversationId: string) {
